# proto/todo.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-example",
      "url": "https://github.com/waveywaves/go-grpc-example"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/todo/all": {
      "get": {
        "summary": "Read all todo tasks",
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/todoReadAllResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ToDoService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "todoPersonal": {
      "type": "object",
      "properties": {
        "conscription": {
          "type": "string"
        },
        "prescription": {
          "type": "string"
        }
      }
    },
    "todoReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/todoToDo"
          },
          "title": "List of all todo tasks"
        }
      },
      "title": "Contains list of all todo tasks"
    },
...
    "todoToDo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Unique integer identifier of the todo task"
        },
        "title": {
          "type": "string",
          "title": "Title of the task"
        },
        "description": {
          "type": "string",
          "title": "Detail description of the todo task"
        },
        "work": {
          "$ref": "#/definitions/todoWork"
        },
        "personal": {
          "$ref": "#/definitions/todoPersonal"
        },
        "reminder": {
          "type": "string",
          "format": "date-time",
          "title": "Date and time to remind the todo task"
        }
      },
      "title": "Task we have to do"
    },
...
    "todoWork": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        }
      }
    }
  }
}